app-backend/requirements.txt
fastapi==0.115.4
passlib==1.7.4
pydantic==2.9.2
python_jose==3.3.0
Requests==2.32.3
SQLAlchemy==2.0.36
uvicorn
psycopg2
bcrypt
python-multipart
logging
